+++ basic_bootloader/src/bootloader/process_transaction.rs
@@ -383,7 +383,7 @@ where
             gas_used,
             gas_refunded: evm_refund,
             computational_native_used,
-            pubdata_used: pubdata_used + L1_TX_INTRINSIC_PUBDATA as u64,
+            pubdata_used: pubdata_used + L1_TX_INTRINSIC_PUBDATA,
         })
     }
 
@@ -734,7 +734,7 @@ where
             gas_used,
             gas_refunded: evm_refund,
             computational_native_used,
-            pubdata_used: pubdata_used + L2_TX_INTRINSIC_PUBDATA as u64,
+            pubdata_used: pubdata_used + L2_TX_INTRINSIC_PUBDATA,
         })
     }
